Extract Groups management from HomeDBManager
